Upload big images to twitter
this is a PITA because of oauth 1, which requires all requests to
through a proxy 'auth-server' to keep the oauth client secret. So when
uploading an image, every byte of data is relayed to twitter via the
auth server. So it takes 2x as long.

(Facebook takes the image as a URL and fetches it herself)